+++ pkg/gohai/verity.go
@@ -6,9 +6,9 @@ import (
 	"os"
 
 	"github.com/DataDog/verity/cpu"
-	"github.com/DataDog/verity/filesystem"
-	"github.com/DataDog/verity/memory"
-	"github.com/DataDog/verity/network"
+	// "github.com/DataDog/verity/filesystem"
+	// "github.com/DataDog/verity/memory"
+	// "github.com/DataDog/verity/network"
 	"github.com/DataDog/verity/platform"
 )
 
@@ -19,9 +19,9 @@ type Collector interface {
 
 var collectors = []Collector{
 	&cpu.Cpu{},
-	&filesystem.FileSystem{},
-	&memory.Memory{},
-	&network.Network{},
+	// &filesystem.FileSystem{},
+	// &memory.Memory{},
+	// &network.Network{},
 	&platform.Platform{},
 }
